requirements.txt
appdirs==1.4.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
codecov==2.0.22
coverage==5.0.4
distlib==0.3.0
Django==2.2.11
djangorestframework==3.10.3
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tox==3.14.5
urllib3==1.25.8
virtualenv==20.0.11
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==3.1.0
